# ornn-frontend/src/types/domain.ts
export interface SkillSummary {
  guid: string;
  name: string;
  description: string;
  createdBy: string;
  createdOn: string;
  isPrivate: boolean;
  tags: string[];
  /** Optional; present when returned from search but not always */
  updatedOn?: string;
}

export interface SkillDetail extends SkillSummary {
  updatedOn: string;
  presignedPackageUrl: string;
  metadata: Record<string, unknown>;
}